requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
multidict==4.7.6
numpy==1.19.1
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
showdownpy==0.1.5
six==1.15.0
typing-extensions==3.7.4.2
urllib3==1.25.10
websockets==8.1
yarl==1.5.0
